Fixed SplitComponent ignoring [size] values
The issue is caused when a `SplitComponent` is added dynamically as a nested split with `[useTransition]="true"`.
When the `SplitAreaDirective` runs it's parent `SplitComponent`'s `addArea` method, the container size is still zero since the transition didn't run yet.

The change causes the `addArea` method to run the `build` method after a timeout equal in length to the transition time (300ms)